+++ tests/unit/test_u_jwe.py
@@ -7,10 +7,19 @@ import jwe
 
 class TestJwe():
     """Tests module jwe.py."""
+    def test_get_wrapped_key_as_jwe(get_jwt):
+        nonce = 'random-nonce'
+        jwe_kid = 'jwe-kid-salesforce-serviceX'
+        jwe_token = jwe.get_wrapped_key_as_jwe(bytearray('randomdek'.encode()),
+                                               'salesforce', jwe_kid, nonce)
+        assert json.loads(jwe_token)['kid'] == jwe_kid
+
     def test__get_key_consumer_cert(self):
-        key_consumer_cert = jwe._get_key_consumer_cert('salesforce', 'jwe-kid-salesforce-serviceX')
+        key_consumer_cert = jwe._get_key_consumer_cert(
+            'salesforce', 'jwe-kid-salesforce-serviceX')
         assert isinstance(key_consumer_cert, str)
-        if isinstance(key_consumer_cert, str) and key_consumer_cert.startswith('-----BEGIN CERTIFICATE-----'):
+        if isinstance(key_consumer_cert, str) and \
+                key_consumer_cert.startswith('-----BEGIN CERTIFICATE-----'):
             assert True, 'Looks like a legitimate certificate.'
         else:
             assert False, 'Cannot detect certificate.'
@@ -35,7 +44,8 @@ class TestJwe():
 
     def test__encrypt_dek_with_cek(self, get_protected_headers):
         b64_dek_ciphertext, b64_tag = jwe._encrypt_dek_with_cek(
-            b'16bitrandom12345', b'randomiv', b'randomdek', get_protected_headers)
+            b'16bitrandom12345', b'randomiv', b'randomdek',
+            get_protected_headers)
 
         assert isinstance(b64_dek_ciphertext, bytes)
         assert isinstance(b64_tag, bytes)
@@ -56,16 +66,20 @@ class TestJwe():
             base64.urlsafe_b64encode(b'tag')
             )
 
-        assert json.loads(json_jwe_token)['kid'] == 'jwe-kid-salesforce-serviceX'
+        assert json.loads(json_jwe_token)['kid'] == \
+            'jwe-kid-salesforce-serviceX'
         assert json.loads(json_jwe_token)['jwe']
-        protected_header = base64.urlsafe_b64decode(json.loads(json_jwe_token)['jwe'].split('.')[0])
-        assert json.loads(protected_header)['kid'] == 'jwe-kid-salesforce-serviceX'
+        protected_header = base64.urlsafe_b64decode(
+            json.loads(json_jwe_token)['jwe'].split('.')[0])
+        assert json.loads(protected_header)['kid'] == \
+            'jwe-kid-salesforce-serviceX'
         assert json.loads(protected_header)['jti'] == 'nonce'
 
     def test__get_jwe_protected_header(self):
         nonce = 'random-nonce'
         jwe_kid = 'jwe-kid-salesforce-serviceX'
         b64_protected_header = jwe._get_jwe_protected_header(jwe_kid, nonce)
-        protected_header = json.loads(base64.urlsafe_b64decode(b64_protected_header))
+        protected_header = json.loads(base64.urlsafe_b64decode(
+            b64_protected_header))
         assert protected_header['kid'] == jwe_kid
         assert protected_header['jti'] == nonce
